+++ app/src/main/java/com/example/wr/story/ui/content/main/MainContract.java
@@ -14,7 +14,7 @@ public interface MainContract{
     interface View extends BaseView {
         void onRecyclerViewAdapterUpdated();
         StorySectionAdapter getRecyclerViewAdapter();
-        void showDetailActivityByStoryId(int storyId);
+        void showDetailActivityByStoryId(long storyId);
     }
 
     interface Presenter {
